Move ConnectionManager to a separate component Move Deluge to a better location
