+++ app/actions/hosts.js
@@ -1,6 +1,6 @@
 // @flow
-import type { Host } from '../api/types'
-import Deluge from '../api/Deluge'
+import type { Host } from '../lib/types'
+import Deluge from '../lib/Deluge/Deluge'
 
 export const HOST_ADDED = 'HOST_ADDED'
 export const REQUEST_ADD_HOST = 'REQUEST_ADD_HOST'
