+++ app/actions/session.js
@@ -1,6 +1,6 @@
 // @flow
-import type { AppState, Torrent } from '../api/types'
-import Deluge from '../api/Deluge'
+import type { AppState, Torrent } from '../lib/types'
+import Deluge from '../lib/Deluge/Deluge'
 
 export const RECEIVE_SESSION_SPEED = 'RECEIVE_SESSION_SPEED'
 export const SHOW_SESSION_SPEED = 'SHOW_SESSION_SPEED'
