+++ app/components/AuthMatch.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Match } from 'react-router'
-import RequireAuth from './RequireAuth'
+import RequireAuth from '../containers/RequireAuth'
 
 const AuthMatch = props => (
   <RequireAuth>
