+++ app/containers/index.js
@@ -5,4 +5,4 @@ export { default as TorrentContainer } from './TorrentContainer'
 export { default as ConnectionManagerContainer } from './ConnectionManagerContainer'
 export { default as TorrentDetailsContainer } from './TorrentDetailsContainer'
 export { default as RequireAuth } from './RequireAuth'
-export { default as AuthMatch } from './AuthMatch'
+export { default as AuthMatch } from '../components/AuthMatch'
