+++ app/lib/Deluge/Auth.js
