+++ app/lib/Deluge/Daemon.js
