+++ app/lib/Deluge/Deluge.js
