+++ app/lib/Deluge/WebUi.js
