+++ app/main.jsx
@@ -5,7 +5,7 @@ import 'react-mdl/extra/material.css'
 import { AppContainer as HMRContainer } from 'react-hot-loader'
 import { render } from 'react-dom'
 import DelugeWUIM from './DelugeWUIM'
-import Deluge from './api/Deluge'
+import Deluge from './lib/Deluge/Deluge'
 import { delugeLocation } from '../settings.json'
 import createAppStore from './store'
 import './main.css'
