+++ app/store.js
@@ -2,8 +2,8 @@
 import { createStore, compose, applyMiddleware } from 'redux'
 import thunk from 'redux-thunk'
 import rootReducer from './reducers'
-import Deluge from './api/Deluge'
-import type { AppState } from './api/types'
+import Deluge from './lib/Deluge/Deluge'
+import type { AppState } from './lib/types'
 
 const defaultState: AppState = {
   torrents: [],
